+++ src/Telegram.Bot/Requests/Payments/SendInvoiceRequest.cs
@@ -36,7 +36,7 @@ public partial class SendInvoiceRequest() : RequestBase<Message>("sendInvoice"),
 
     /// <summary>Identifier of the direct messages topic to which the message will be sent; required if the message is sent to a direct messages chat</summary>
     [JsonPropertyName("direct_messages_topic_id")]
-    public int? DirectMessagesTopicId { get; set; }
+    public long? DirectMessagesTopicId { get; set; }
 
     /// <summary>Payment provider token, obtained via <a href="https://t.me/botfather">@BotFather</a>. Pass an empty string for payments in <a href="https://t.me/BotNews/90">Telegram Stars</a>.</summary>
     [JsonPropertyName("provider_token")]
