+++ src/Telegram.Bot/Types/PaidMediaInfo.cs
@@ -7,7 +7,7 @@ public partial class PaidMediaInfo
     /// <summary>The number of Telegram Stars that must be paid to buy access to the media</summary>
     [JsonPropertyName("star_count")]
     [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
-    public int StarCount { get; set; }
+    public long StarCount { get; set; }
 
     /// <summary>Information about the paid media</summary>
     [JsonPropertyName("paid_media")]
